Move `NavigatorScreen` under `navigator-screen` folder, add README
