+++ packages/components/src/navigator/index.js
@@ -1,3 +1,3 @@
 export { default as NavigatorRoot } from './navigator-root';
-export { default as NavigatorScreen } from './screen';
+export { default as NavigatorScreen } from './navigator-screen';
 export { default as useNavigator } from './use-navigator';
