+++ packages/components/src/navigator/navigator-screen/README.md
@@ -1,0 +1,22 @@
+# `NavigatorScreen`
+
+<div class="callout callout-alert">
+This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
+</div>
+
+The `NavigatorScreen` component represents a single view/screen/panel/menu and is supposed to be used in combination with [the `NavigatorRoot` component](/packages/components/src/navigator/navigator-root/README.md).
+
+## Usage
+
+Refer to [the `NavigatorRoot` component](/packages/components/src/navigator/navigator-root/README.md#usage) for a usage example.
+
+## Props
+
+The component accepts the following props:
+
+### `path`
+
+-   Type: `string`
+-   Required: Yes
+
+The path of the current screen.
