+++ packages/components/src/navigator/navigator-screen/index.js
@@ -1,0 +1,1 @@
+export { default } from './component';
